RajeevB_Reading1.md
In Yvette Granata's article, she explained about the given name to any different type of fish called the Lampreys. In addition, she discussed how that parasitic jawless fish engulf the blood of its host. 
Moreover, she added that according to Google and her studies on those creatures, she found pictures of lampreys that has never been seen before. Inexplicable to explain, those pictures have revealed many features of the lampreys' mouth as well as its teeth with the help of nowadays best technology, the latest high definition cameras. Those high-quality images from Google that Yvette found of the lampreys is are images of an animal body and of its sharp teeth. She also mentions about a photograph of an animal-object as a parasite that strikes to latch upon the viewer.
Furthermore, she stated concerning that we may also view Google’s image search as reinforcing this idea, replicating a similar graphic processing of images within the internal prerogative of its exploration algorithm that really intrigued my attention at some point. Staring at those pictures, especially the ones where they focused on the mouth made me think of the "Sarlacc Pit" in the movie Star Wars. Besides, Yvette talks about the Processing Lampreys project produces the digital process of coding as a photographic means, it begins with an allocation of lamprey images from the google search results. In addition to that, she mentions how abstract stretchings and movements of the lamprey picture, turning the shot of the lamprey into a digital array. Eventually, to her own opinion, she mentions that "with digital image processing, we are more like the lamprey."